added ndk reference, now I can use ndk. Added Tess.
the reference to ndk is in local.properties
lots of things were added in the build.gradle.
including:  module name of ndk module to be built
            product flavors (only the x86 one is necessary
            application variants have specific versions

Added tesseract and leptonica libs. TODO: ocrfragment
that passes intent to camera app and recieves the
image response.. also, a share image with app
should also send the image to the ocrfragment, except
that the fragment should run ocr instead of opening
camera intent.

Note: Not sure yet if user should see the
results of ocr's boxes over the image? or if they should
simply see the text result.. Leaning on the latter..
If thats the case then simply run spritz if text is valid
after ocr-ing.