Added epublib. Created apub-fragment.
When an epub from the filesystem is shared or sent
to the app then the main activity will send the book
to the epub fragment, load the epub fragment and the
epub fragment will be used to show the epub.

The user should be able to set a cursor in the text
to mark the position from where to start spritzing.
(this is the alternative to simply spritzing the text
in the epub right away without any user input on where
to start spritzing..)

TODO: (1) extract text from book.
      (2) view pages of text in epub fragment
      (3) cursor on text and play to spritz button
      (4) open spritz fragment with some "lazy loaded"
          text or pages.. is there a a Spritz source
          compatible with epublib resource objects.