Built non-ocr apk. Added epub shared preferences.
Shared preferences: Pagesize, Fontsize, CSS Selector

Page size is the number of html lines to extract from
the epub per webview 'page'.. Updated this extraction
method to include only the html elements described by
the selector and with inner text content.

Fontsize is the size of the font in the webview

CSS Selector is used to select elements from the epub
html to paste onto the webview 'page'. It is also the
same selector used to select the elements and extract
text from the page for spritzing.