Added context menu per Text Partition Item. Fixed preferences.
Fixed preferences by setting default colors for some spritz options.
addec context menu, edit and delete capabilities. This is temporary..
It should be one of those actionbar context menus with delete bin..
and it should also allow text edit in place.. however the benefit of
the edit text being in the dialog is that its easier to see the text
because it takes up more of the screen.. Maybe a cool magnification
animation that magnifies the text into a bubble view..